Fix cache for LD instructions

+++ src/mu-mips.c
@@ -2532,8 +2532,8 @@ void show_pipeline(){
 	printf("MEM_WB.RegisterRd:%d\n\n", MEM_WB.RegisterRd);
 	printf("CYCLE %u\n", CYCLE_COUNT);
 	int total_accesses = cache_hits + cache_misses;
-	double hit_rate = cache_hits / total_accesses;
-	double miss_rate = 1 - hit_rate;
+	double hit_rate = (double)cache_hits / (double)total_accesses;
+	double miss_rate = 1.00 - hit_rate;
 	printf("Hit rate: %lf%%\n", hit_rate*100);
 	printf("Miss rate: %lf%%\n", miss_rate*100);
 	printf("Total Hit: %d\n Total Miss: %d\n Total Accesses: %d\n", cache_hits, cache_misses, total_accesses);
